step sopra utils e main
